chore(all): release v0.0.11

+++ cssbattle-web/src/utils/request.ts
@@ -4,7 +4,7 @@
  * @Author: pym
  * @Date: 2021-08-29 19:33:21
  * @LastEditors: pym
- * @LastEditTime: 2021-10-07 21:49:14
+ * @LastEditTime: 2021-10-10 12:42:52
  */
 import { message } from 'antd';
 import axios from 'axios';
@@ -18,7 +18,9 @@ axios.defaults.validateStatus = function (status: any) {
 //跨域请求，允许保存cookie
 axios.defaults.withCredentials = true;
 axios.interceptors.request.use((config: any) => {
-  config.headers['token'] = localStorage.getItem('token')
+  if(localStorage.getItem('token')) {
+    config.headers['token'] = localStorage.getItem('token')
+  }
   if(!config.url){
     return Promise.reject(new Error())
   }
@@ -31,15 +33,14 @@ axios.interceptors.request.use((config: any) => {
 
 //HTTPrequest拦截
 axios.interceptors.response.use((config: any) => {
-  console.log(config)
-  if(config.status === 401) {
+  if(config && config.status === 401) {
     localStorage.clear()
     window.location.href = `${location.origin}/index/login`
   }
   return config;
 }, (error: any) => {
-  message.error('服务请求失败！');
-  return Promise.reject(new Error(error));
+  message.error(error.response.data.msg || '服务请求失败');
+  return Promise.reject();
 })
 
 export default axios;
